Added permissions to the project detail page

+++ src/pages/api/projects/[id]/index.ts
@@ -1,28 +1,75 @@
+import { permissions } from '../../../../config';
 import {
 	prisma,
 	getProject,
 	projectSelectQuery as selectQuery,
 } from '../../../../db';
+import {
+	getRecord,
+	getUserObjectPermissions,
+	removeObjectPermissions,
+	updateObjectPermissions,
+} from '../../../../db/utils';
 import { auth } from '../../../../middlewares';
-import { CreateProjectQueryType } from '../../../../types';
+import { adminMiddleware as admin } from '../../../../middlewares/api';
+import { CreateProjectQueryType, ProjectType } from '../../../../types';
+import { hasModelPermission } from '../../../../utils';
+import { NextApiErrorMessage } from '../../../../utils/classes';
 import { projectCreateSchema } from '../../../../validators';
 
 export default auth()
 	.get(async (req, res) => {
-		const project = await getProject(req.query.id as string);
-		if (!project) {
+		const record = await getRecord({
+			model: 'projects',
+			perm: 'project',
+			permission: 'VIEW',
+			objectId: req.query.id as string,
+			user: req.user,
+			getData() {
+				return getProject(req.query.id as string);
+			},
+		});
+
+		if (!record) throw new NextApiErrorMessage(403);
+
+		if (!record.data) {
 			return res.status(404).json({
 				status: 'error',
 				message: 'Project with specified ID does not exist',
 			});
 		}
+
 		return res.status(200).json({
 			status: 'success',
 			mesage: 'Fetched project successfully!',
-			data: project,
+			data: record.data,
 		});
 	})
+	.use(admin)
 	.put(async (req, res) => {
+		let hasPerm =
+			req.user.isSuperUser ||
+			hasModelPermission(req.user.allPermissions, [permissions.project.EDIT]);
+
+		if (!hasPerm) {
+			const perm = await getUserObjectPermissions({
+				modelName: 'projects',
+				permission: 'EDIT',
+				objectId: req.query.id as string,
+				userId: req.user.id,
+			});
+			hasPerm = perm.edit;
+		}
+
+		if (!hasPerm) throw new NextApiErrorMessage(403);
+
+		const previousProject = await getProject(req.query.id as string);
+		if (!previousProject)
+			return res.status(404).json({
+				status: 'error',
+				message: 'Project with the specified ID does not exist',
+			});
+
 		const valid: CreateProjectQueryType =
 			await projectCreateSchema.validateAsync({ ...req.body });
 
@@ -35,7 +82,7 @@ export default auth()
 			});
 
 		// update the project
-		const project = await prisma.project.update({
+		const project = (await prisma.project.update({
 			where: { id: req.query.id as string },
 			data: {
 				...data,
@@ -44,6 +91,10 @@ export default auth()
 							connect: {
 								id: data.client,
 							},
+					  }
+					: previousProject.client
+					? {
+							disconnect: true,
 					  }
 					: undefined,
 				team:
@@ -60,6 +111,36 @@ export default auth()
 						: {},
 			},
 			select: selectQuery,
+		})) as unknown as ProjectType;
+
+		if (
+			(!data.client && previousProject.client) ||
+			(previousProject.client &&
+				data.client &&
+				data.client !== previousProject.client?.id)
+		) {
+			// Remove the old client object permissions
+			await removeObjectPermissions({
+				model: 'projects',
+				objectId: project.id,
+				users: [previousProject.client.contact.id],
+			});
+		}
+
+		const viewers: string[] = [];
+		if (project.client) {
+			viewers.push(project.client.contact.id);
+		}
+
+		project.team.forEach((member) => {
+			viewers.push(member.employee.user.id);
+		});
+
+		await updateObjectPermissions({
+			model: 'projects',
+			permissions: ['VIEW'],
+			objectId: project.id,
+			users: viewers,
 		});
 
 		return res.status(200).json({
@@ -69,6 +150,22 @@ export default auth()
 		});
 	})
 	.delete(async (req, res) => {
+		let hasPerm =
+			req.user.isSuperUser ||
+			hasModelPermission(req.user.allPermissions, [permissions.project.DELETE]);
+
+		if (!hasPerm) {
+			const perm = await getUserObjectPermissions({
+				modelName: 'projects',
+				permission: 'VIEW',
+				objectId: req.query.id as string,
+				userId: req.user.id,
+			});
+			hasPerm = perm.delete;
+		}
+
+		if (!hasPerm) throw new NextApiErrorMessage(403);
+
 		await prisma.project.delete({ where: { id: req.query.id as string } });
 		return res.status(200).json({
 			status: 'success',
